Support for returning futures.Future objects transparently
Shortly, a request that returns a Future object, will return
a dummy (proxy-side) Future object linked to the original
(server-side) Future object via the PUB/SUB socket.

When a Proxy calls a method (or request an attribute)
in the Server that returns a Future object:

 1.- A callback is added to the Future (at the Server)
 2.- A message is sent back to the proxy to acknowledge
     the request. The proxy will create dummy Future object,
     start it and store in dictionary using an identification
     provided by the server as key. (This identification is
     currently the msgid of the Request)
     The proxy return de Future object

Once that the future object is completed, the callback
on the Server will be called. This callback will notify
the Proxy using the PUB/SUB connection about the result
and any exceptions that had been raised. The Proxy will then
call the set_result and set_exception method.

Not supported yet: Future.cancel()