Mechanism for a more efficient Proxy Server communication
When the proxy connects to the server, it will request a
set of attributes that will be returned as RemoteAttribute
and a set of attributes that will be returned as objects.
The set of RemoteAttributes has precedence over the ones
of objects.
Attributes not mentioned in these sets will follow the
usual path.

A class derived from Server could reimplement `inspect`,
`return_as_remote` or `force_as_object` for specific
functionality